add/ refining the radar chart and connecting to goal input
Prompt: " I currently have GoalTracker and RadarChart but they aren't interacting. I want to change it so that when a goal is entered into the goal tracker, it's added as a label with a value into the desired skills dataset item." > applied suggestions with no changes

Prompt: I want to be able to provide the value for the new skill as well. Can you help me change the goal input to make it clear how to provide that along with the skill name? > applied suggestions with no changes

Prompt: Currently this is adding to the desired skills, which is correct. I want that to stay the same. But, I also want to be able to supply a starting value when adding a new skill. How can I do that? > applied suggestions with no changes

Also addressed dependancy issue: @babel/plugin-proposal-private-property-in-object